make GT++ CommonProxy implement IFuelHandler directly instead of using a separate class
